src/function_app/requirements.txt
azure-functions
azure-communication-email>=1.0.0
pandas>=2.0.0
